Clean: Removed unused Variables, Organized packages
