+++ Planner/app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name="rdatu.android.cyscorpions.com.projectplanner.view.HomeActivity"
+            android:name=".view.activities_fragments.HomeActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -17,13 +17,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="rdatu.android.cyscorpions.com.projectplanner.view.ListPlannerActivity"
+            android:name=".view.activities_fragments.ListPlannerActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait">
 
         </activity>
         <activity
-            android:name="rdatu.android.cyscorpions.com.projectplanner.view.ScheduleTaskActivity"
+            android:name=".view.activities_fragments.ScheduleTaskActivity"
             android:label="ProjectPlanner"
             android:screenOrientation="portrait">
