+++ Planner/app/src/main/java/rdatu/android/cyscorpions/com/projectplanner/controller/adapters/PlannerPagerAdapter.java
@@ -1,10 +1,10 @@
-package rdatu.android.cyscorpions.com.projectplanner.controller;
+package rdatu.android.cyscorpions.com.projectplanner.controller.adapters;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
 
-import rdatu.android.cyscorpions.com.projectplanner.view.ListPlannerFragment;
+import rdatu.android.cyscorpions.com.projectplanner.view.activities_fragments.ListPlannerFragment;
 
 /**
  * Created by rayeldatu on 7/27/15.
