+++ Planner/app/src/main/java/rdatu/android/cyscorpions/com/projectplanner/view/activities_fragments/ListPlannerFragment.java
@@ -1,4 +1,4 @@
-package rdatu.android.cyscorpions.com.projectplanner.view;
+package rdatu.android.cyscorpions.com.projectplanner.view.activities_fragments;
 
 
 import android.app.Activity;
@@ -25,8 +25,9 @@ import java.util.ArrayList;
 import java.util.Calendar;
 
 import rdatu.android.cyscorpions.com.projectplanner.R;
-import rdatu.android.cyscorpions.com.projectplanner.controller.TaskManager;
-import rdatu.android.cyscorpions.com.projectplanner.model.Tasks;
+import rdatu.android.cyscorpions.com.projectplanner.controller.modelmanagers.TaskManager;
+import rdatu.android.cyscorpions.com.projectplanner.model.objects.Tasks;
+import rdatu.android.cyscorpions.com.projectplanner.view.dialogs.DatePickerDialog;
 
 /**
  * Created by rayeldatu on 7/27/15.
