+++ Planner/app/src/main/java/rdatu/android/cyscorpions/com/projectplanner/view/dialogs/TimePickerDialog.java
@@ -1,4 +1,4 @@
-package rdatu.android.cyscorpions.com.projectplanner.view;
+package rdatu.android.cyscorpions.com.projectplanner.view.dialogs;
 
 import android.app.Activity;
 import android.app.Dialog;
